feat: update types
