+++ src/types/TodoType.ts
@@ -1,0 +1,5 @@
+export interface TodoType {
+  id: number;
+  title: string;
+  completed: boolean;
+}
